add pinnedSheet to threadPage
